requirements.txt
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
scipy==1.15.2
